Fix #15, #16
Issue #16: RTeasy still hangs on the line but now an error message is
displayed in the log. Therefore step and microstep methods in RTProgram are
adjusted. They now do a cycle change only of no error occurs. RTSimWindow
is also adjusted to display errors in the log.

Issue #15: Statements after if are evaluated now. Therefore the
getStatementsOrdered method in ParallelStatements is adjusted to add if
statements at last to the output.